(FMT): simplify
