+++ src/main/kotlin/org/rust/ide/formatter/RustFormattingModelBuilder.kt
@@ -9,14 +9,14 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.codeStyle.CodeStyleSettings
-import org.rust.ide.formatter.blocks.AbstractRustFmtBlock
+import org.rust.ide.formatter.blocks.RustFmtBlock
 
 class RustFormattingModelBuilder : FormattingModelBuilder {
     override fun getRangeAffectingIndent(file: PsiFile?, offset: Int, elementAtOffset: ASTNode?): TextRange? = null
 
     override fun createModel(element: PsiElement, settings: CodeStyleSettings): FormattingModel {
         val ctx = RustFmtContext.create(settings)
-        val block = AbstractRustFmtBlock.createBlock(element.node, null, Indent.getNoneIndent(), null, ctx)
+        val block = RustFmtBlock.createBlock(element.node, null, Indent.getNoneIndent(), null, ctx)
         /** / com.intellij.formatting.FormattingModelDumper.dumpFormattingModel(block, 2, System.err) // */
         return FormattingModelProvider.createFormattingModelForPsiFile(element.containingFile, block, settings)
     }
